Install llvm-mingw
This bundles clang/lld and associated tools (configured for a mingw
target) with corresponding mingw sysroots.

As we already install the 'llvm' chocolatey package, which also
provides clang (defaulting to a MSVC target), add this one last in the
path, and refer to with the GCC style triple prefixed driver names,
such as x86_64-w64-mingw32-clang.